+++ app/src/test/java/eu/kanade/tachiyomi/data/backup/SourcePreferencesBackupTest.kt
@@ -101,6 +101,7 @@ class SourcePreferencesBackupTest {
             novelPreferences = mockk(),
             extensionSourcePreferences = mockk(),
             networkPreferences = mockk(),
+            getNovelCategories = mockk(),
         )
 
         restorer.restoreSource(